% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addressApiDownload.R
\name{addressApiDownload}
\alias{addressApiDownload}
\title{Downloads the case API from Bioportal on the selected date range}
\usage{
addressApiDownload(email, password, format = FALSE)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}

\item{format}{boolian variable. If true, returns the formatted data frame}
}
\value{
data frame of all tests starting at date from  and ending at date to of the unique tests API
}
\description{
Downloads the case API from Bioportal on the selected date range
}
\examples{
testApiDownloadAll(email, password)
Datos_all=addressApiDownloadAll(my_email, my_password)
---format defaults to FALSE, as format function not yet implemented for cases ---
}
