% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_Awesome.R
\name{BP_Awesome}
\alias{BP_Awesome}
\title{Prints BioPortal is awesome :)}
\usage{
BP_Awesome()
}
\value{
The same Data Frame with date columns transformed into R Date objects (%Y-%m-%d)
}
\description{
Prints BioPortal is awesome :)
}
