% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseApiDownloadAll.R
\name{caseApiDownloadAll}
\alias{caseApiDownloadAll}
\title{Downloads the enteirity of the unique case API}
\usage{
caseApiDownloadAll(email, password)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}
}
\value{
(very large!!) data frame of all cases from case API
}
\description{
Downloads the enteirity of the unique case API
}
\examples{
caseApiDownloadAll(email, password)
Datos_all=caseApiDownloadAll(my_email, my_password)
}
