% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseApiGetAddressDataFrame.R
\name{caseApiGetAddressDataFrame}
\alias{caseApiGetAddressDataFrame}
\title{extracts the address as a dataframe}
\usage{
caseApiGetAddressDataFrame(caseDataFrame)
}
\arguments{
\item{caseDataFrame}{data frame of entities obtained from the entity API}
}
\value{
data frame of cases with case ids, number of addresses and info on missing values
}
\description{
extracts the address as a dataframe
}
\examples{
caseApiAddressDataFrame=caseApiGetAddressDataFrame(caseDataFrame)

}
