% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityApiDownloadAll.R
\name{entityApiDownloadAll}
\alias{entityApiDownloadAll}
\title{Downloads the enteirity of the unique case API}
\usage{
entityApiDownloadAll(email, password)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}
}
\value{
data frame of all entities from entity API
}
\description{
Downloads the enteirity of the unique case API
}
\examples{
entityApiDownloadAll(email, password)
Entities=entityApiDownloadAll(my_email, my_password)
}
