% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperExtractCliaFromEntity.R
\name{helperExtractCliaFromEntity}
\alias{helperExtractCliaFromEntity}
\title{extracts the clia}
\usage{
helperExtractCliaFromEntity(labEntity)
}
\arguments{
\item{labEntity}{a list of identifiers for a given row of the entityApi data frame}
}
\value{
the string stored under clia, if any. If no clia is found it returns the empty string
}
\description{
extracts the clia
}
\examples{

}
