% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperMonthDelimiter.R
\name{helperMonthDelimiter}
\alias{helperMonthDelimiter}
\title{generates two vectors, one with the first day of the month starting from april 2020
to the current month, and one with le last day of the month starting from april 2020
to the current month}
\usage{
helperMonthDelimiter(date = Sys.Date())
}
\arguments{
\item{date}{A date or string to be interpreted as.Date()}
}
\value{
list with two vectors, one for the start of month and one for the end of month
}
\description{
generates two vectors, one with the first day of the month starting from april 2020
to the current month, and one with le last day of the month starting from april 2020
to the current month
}
