% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperSliceByMonth.R
\name{helperSliceByMonth}
\alias{helperSliceByMonth}
\title{returns a list of dates at most a month apart to use for downloading data}
\usage{
helperSliceByMonth(from, to)
}
\arguments{
\item{from}{A date or string from which to start the sequence}

\item{to}{A date or string from which to end the sequence}
}
\value{
list with two vectors, starting at from and ending at to, sliced by month
}
\description{
returns a list of dates at most a month apart to use for downloading data
}
\examples{
helperSliceByMonth("2021-01-15","2021-01-26")
helperSliceByMonth("2021-01-15","2021-02-27")

}
