% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalizationsApiDownloadAll.R
\name{hospitalizationsApiDownloadAll}
\alias{hospitalizationsApiDownloadAll}
\title{Downloads the enteirity of the hospitalizations API}
\usage{
hospitalizationsApiDownloadAll(email, password)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}
}
\value{
data frame of all entities from entity API
}
\description{
Downloads the enteirity of the hospitalizations API
}
\examples{
hospitalizationsApiDownloadAll(email, password)
patients=hospitalizationsApiDownloadAll(my_email, my_password)
}
