% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LabTags}
\alias{LabTags}
\title{LabTags Data Set}
\format{
A data frame with 322 rows and 3 variables:
\describe{
  \item{\code{Nombre}}{character Bioportal Name}
  \item{\code{ID}}{character SARAF ID}
  \item{\code{Tag}}{character General Tag}
}
}
\usage{
LabTags
}
\description{
Holds bioportal name, ids and general tag
}
\keyword{datasets}
