% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagApiAddLIS.R
\name{lagApiAddLIS}
\alias{lagApiAddLIS}
\title{Adds a colummn with the LIS to the 'lag' API DataFrame}
\usage{
lagApiAddLIS(LagDataFrame)
}
\arguments{
\item{LagDataFrame}{A Data Frame object from lag API}
}
\value{
DataFrame with an additional colummn containing the LIS to the 'lag' API DataFrame
}
\description{
Adds a colummn with the LIS to the 'lag' API DataFrame
}
