% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagApiBizDelay.R
\name{lagApiBizDelay}
\alias{lagApiBizDelay}
\title{calculates the delays between file upload and created, considering the number of buisness days}
\usage{
lagApiBizDelay(LagDataFrame)
}
\arguments{
\item{LagDataFrame}{A Data Frame object from lag API}
}
\value{
DataFrame with an additional colummn containing the delays between file upload and created, considering the number of buisness days
}
\description{
calculates the delays between file upload and created, considering the number of buisness days
}
