% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagApiCalculateTurnaround.R
\name{lagApiCalculateTurnaround}
\alias{lagApiCalculateTurnaround}
\title{Calculates Turnaround time (Tat) for when the sample was reported, when the repost was uploaded and when the sample was uploaded}
\usage{
lagApiCalculateTurnaround(DataFrame)
}
\arguments{
\item{DataFrame}{A Data Frame object from lag API}
}
\value{
The same Data Frame with added Tat columns for sample was reported, when the repost was uploaded and when the sample was uploaded
}
\description{
Calculates Turnaround time (Tat) for when the sample was reported, when the repost was uploaded and when the sample was uploaded
}
\examples{
lagApiCalculateTurnaround(All_Test)
}
