% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagApiFormat.R
\name{lagApiFormat}
\alias{lagApiFormat}
\title{Formats the lag api data frame to have date objects, turnaround times, adds the lab tags, etc.}
\usage{
lagApiFormat(LagDataFrame)
}
\arguments{
\item{LagDataFrame}{A Data Frame object from Test API}
}
\value{
The same Data Frame with date columns transformed in to date objects
}
\description{
Formats the lag api data frame to have date objects, turnaround times, adds the lab tags, etc.
}
