% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagApiTestVolume.R
\name{lagApiTestVolume}
\alias{lagApiTestVolume}
\title{Generates Plots that include dates in the x axis}
\usage{
lagApiTestVolume(
  data_frame = All_Tests,
  cutoff_lower = "no",
  cutoff_upper = "no",
  cutoff_min = FALSE,
  cutoff_max = FALSE
)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}

\item{cutoff_lower}{a date from which to cut off the lower part of the data frame.
displays tests only from cutoff lower onward (as.Date(), defaults to string "no")}

\item{cutoff_upper}{a date from which to cut off the upper part of the data frame.
displays tests only from cutoff lower and earlier (as.Date(), defaults to string "no")}

\item{cutoff_min}{a boolean value. If TRUE, shows the data frame from the earliest non-zero date onward
defaults to FALSE}

\item{cutoff_max}{a boolean value. If TRUE, shows the data frame up to the latest non-zero date onward
defaults to FALSE}
}
\value{
Plot of data changing over time
}
\description{
Generates Plots that include dates in the x axis
}
\examples{
lagApiTestVolume()
lagApiSubset(lab_file="quest", test_type="pcr") \%>\% lagApiTestVolume()

}
