% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsApiDownloadAll.R
\name{patientsApiDownloadAll}
\alias{patientsApiDownloadAll}
\title{Downloads the enteirity of the patients API}
\usage{
patientsApiDownloadAll(email, password, format = FALSE)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}

\item{format}{boolian variable. If true, returns the formatted data frame}
}
\value{
(very large!!) data frame of all tests from unique tests API
}
\description{
Downloads the enteirity of the patients API
}
\examples{
testApiDownloadAll(email, password)
Datos_all=testApiDownloadAll(my_email, my_password)
}
