% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_test_api_analysis.R
\name{save_test_api_analysis}
\alias{save_test_api_analysis}
\title{Saves results from analysis, be it in table or graphic form}
\usage{
save_test_api_analysis(
  AnalysisVar = NULL,
  GraphVar = NULL,
  folder_name,
  date_tag = "today",
  lab_tag,
  file_type_tag,
  saveFile = TRUE,
  saveGraph = TRUE
)
}
\arguments{
\item{AnalysisVar}{Variable that holds table with analysis done to the data, like turn over or volume}

\item{GraphVar}{Variable that holds the graph done for analysis}

\item{folder_name}{Name of the directory you want to create to save your data}

\item{date_tag}{default is set to today but can be changed by writing a date in the format "%Y_%m_%d"}

\item{lab_tag}{Name of the laboratory whose data you are analyzing}

\item{file_type_tag}{character string of type of analysis done}

\item{saveFile}{Do you want to save your table in to a file? default is true}

\item{saveGraph}{Do yo want to save your graph in to an image? default is true}
}
\value{
graph or table of analysis done
}
\description{
Saves results from analysis, be it in table or graphic form
}
\examples{
**work in progress**
}
