% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolMemebersApiDownloadAll.R
\name{schoolMemebersApiDownloadAll}
\alias{schoolMemebersApiDownloadAll}
\title{Downloads the enteirity of the School Memebers API}
\usage{
schoolMemebersApiDownloadAll(email, password)
}
\arguments{
\item{email}{Bioportal account email to be used to access unique tests API}

\item{password}{Bioportal account password to be used to access unique tests API}
}
\value{
data frame of the school Members API
}
\description{
Downloads the enteirity of the School Memebers API
}
\examples{
testApiDownloadAll(email, password)
Datos_all=testApiDownloadAll(my_email, my_password)
}
