% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiAddClia.R
\name{testApiAddLabClia}
\alias{testApiAddLabClia}
\title{Adds two columns with the clia  for orderedByEntity and processedByEntity}
\usage{
testApiAddLabClia(
  testDataFrame,
  cliaDataFrame = NULL,
  email = "my_email",
  password = "my_password"
)
}
\arguments{
\item{testDataFrame}{A Data Frame object from Test API}

\item{cliaDataFrame}{A Data Frame from the entity API (if none is found, it downloads it)}

\item{email}{email used to download the entities dataframe (if needed)}

\item{password}{password used to download the entities dataframe (if needed)}
}
\value{
The same Data Frame with an added Column containg the corresponding Lab Tags for each test
}
\description{
Adds two columns with the clia  for orderedByEntity and processedByEntity
}
\examples{
testApiAddLabClias
testApiAddLabClias(sub_Test)
}
