% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiAddLabTag.R, R/testApiAddTag.R
\name{testApiAddLabTag}
\alias{testApiAddLabTag}
\title{Adds a Column with the Lab Tags}
\usage{
testApiAddLabTag(DataFrame = All_Tests)

testApiAddLabTag(DataFrame = All_Tests)
}
\arguments{
\item{DataFrame}{A Data Frame object from Test API}
}
\value{
The same Data Frame with an added Column containg the corresponding Lab Tags for each test

The same Data Frame with an added Column containg the corresponding Lab Tags for each test
}
\description{
Adds a Column with the Lab Tags

Adds a Column with the Lab Tags
}
\examples{
testApiCalculateTurnaround(All_Test)
testApiCalculateTurnaround(sub_Test)
testApiCalculateTurnaround(All_Test)
testApiCalculateTurnaround(sub_Test)
}
