% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiAddTag.R
\name{testApiAddTag}
\alias{testApiAddTag}
\title{Adds a Column with the Lab Tags}
\usage{
testApiAddTag(DataFrame = All_Tests)
}
\arguments{
\item{DataFrame}{A Data Frame object from Test API}
}
\value{
The same Data Frame with an added Column containg the corresponding Lab Tags for each test
}
\description{
Adds a Column with the Lab Tags
}
\examples{
testApiCalculateTurnaround(All_Test)
testApiCalculateTurnaround(sub_Test)
}
