% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiCalculateTurnaround.R
\name{testApiCalculateTurnaround}
\alias{testApiCalculateTurnaround}
\title{Calculates Turnaround time (Tat) for when the sample was reported, when the repost was uploaded and when the sample was uploaded}
\usage{
testApiCalculateTurnaround(DataFrame = All_Tests)
}
\arguments{
\item{DataFrame}{A Data Frame object from Test API}
}
\value{
The same Data Frame with added Tat columns for sample was reported, when the repost was uploaded and when the sample was uploaded
}
\description{
Calculates Turnaround time (Tat) for when the sample was reported, when the repost was uploaded and when the sample was uploaded
}
\examples{
testApiCalculateTurnaround(All_Test)
testApiCalculateTurnaround(sub_Test)
}
