% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiCountDelays.R
\name{testApiCountDelays}
\alias{testApiCountDelays}
\title{Generates Tables (3) with number of tests by given delay}
\usage{
testApiCountDelays(DataFrame = All_Tests)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by
testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}
}
\value{
Three tables with information on the number of tests by difference in time stamps.
The first one returns the Turnaround Time (TaT) between the report date and the sample date.
The second one returns the Turnaround Time (TaT) between the upload date and the report date.
The third one returns the Turnaround Time (TaT) between the upload date and the sample date.
}
\description{
Generates Tables (3) with number of tests by given delay
}
\examples{
testApiCountDelays()
testApiSubset(test_type="pcr") \%>\% testApiCountLabsAndType()[[2]]

}
