% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiCountLabs.R
\name{testApiCountLabs}
\alias{testApiCountLabs}
\title{Generates Table with number of tests by ordering entity and processing entity (i.e. by lab)}
\usage{
testApiCountLabs(data_frame = All_Tests)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by
testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}
}
\value{
A table with information on the number of tests by lab
}
\description{
Generates Table with number of tests by ordering entity and processing entity (i.e. by lab)
}
\examples{
testApiCountLabs
testApiSubset(test_type="pcr") \%>\% testApiCountLabs()

}
