% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiCountType.R
\name{testApiCountType}
\alias{testApiCountType}
\title{Generates Table with number of tests by type, result and upload method}
\usage{
testApiCountType(data_frame = All_Tests)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by
testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}
}
\value{
A table with information on the number of tests by type, result and upload method
}
\description{
Generates Table with number of tests by type, result and upload method
}
\examples{
testApiCountType
testApiSubset(test_type="pcr") \%>\% testApiCountType()

}
