% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiDatePlot.R
\name{testApiDatePlot}
\alias{testApiDatePlot}
\title{Generates Plots that include dates in the x axis}
\usage{
testApiDatePlot(
  data_frame,
  col = "Sampled",
  colour = "blue",
  x_label = "Fecha",
  y_label = "Cantidad de Pruebas",
  title = "Pruebas",
  infer_labels = FALSE
)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}

\item{col}{"Sampled" colummn to plot (can be "Sampled", "Reported", or "Uploaded")}

\item{colour}{"blue" color to use for fill}

\item{y_label}{"Cantidad de Pruebas"}

\item{title}{"Pruebas"}

\item{infer_labels}{FALSE}
}
\value{
Plot of data changing over time
}
\description{
Generates Plots that include dates in the x axis
}
\examples{
testApiDatePlot(PCR_Test)
testApiDatePlot(PCR_Test, col = "Reported")

}
