% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiDownload.R
\name{testApiDownload}
\alias{testApiDownload}
\title{Downloads the enteirity of the unique tests API}
\usage{
testApiDownload(email, password, from, to, format = TRUE)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}

\item{from}{Date (upload Date) from where to start downloading data.}

\item{to}{Date (upload date) to where to download the data (the last date is included)}

\item{format}{boolian variable. If true, returns the formatted data frame}
}
\value{
data frame of all tests starting at date from  and ending at date to of the unique tests API
}
\description{
Downloads the enteirity of the unique tests API
}
\examples{
testApiDownloadAll(email, password, from, to)
Datos_all=testApiDownloadAll(my_email, my_password,"2021-01-01", "2021-03-15")
}
