% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiETR.R
\name{testApiETR}
\alias{testApiETR}
\title{Returns Positivity for test available in TestApi Table}
\usage{
testApiETR(
  DataFrame = All_Tests,
  cutoff_min = TRUE,
  cutoff_max = TRUE,
  cutoff_lower = "no",
  cutoff_upper = "no"
)
}
\arguments{
\item{DataFrame}{A data frame that has already been manipulated by testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}

\item{cutoff_min}{If True, cuts off to the earliest non-zero date}

\item{cutoff_max}{If True, cuts off to the latest non-zero date}

\item{cutoff_lower}{Cuts off at specific lower date}

\item{cutoff_upper}{Cuts off at specific upper date}
}
\value{
NewFrame with the Electronic Transmission Rate (ETR) calculated by date
}
\description{
Returns Positivity for test available in TestApi Table
}
\examples{
testApiPositivity(PCR_Test)
testApiPositivity(ByLab_Test)

}
