% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiETRPlot.R
\name{testApiETRPlot}
\alias{testApiETRPlot}
\title{Plots ETR results}
\usage{
testApiETRPlot(DataFrame, fromTests = FALSE)
}
\arguments{
\item{DataFrame}{A Data Frame object with calculated ETR}
}
\value{
ETR Plot
}
\description{
Plots ETR results
}
\examples{
testApiETR(All_Tests)\%>\%testApiPositivityPlot()
}
