% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiFormat.R
\name{testApiFormat}
\alias{testApiFormat}
\title{Formats the test api data frame to have date objects, turnaround times, adds the lab tags, etc.}
\usage{
testApiFormat(dataFrame)
}
\arguments{
\item{DataFrame}{A Data Frame object from Test API}
}
\value{
The same Data Frame with date columns transformed in to date objects
}
\description{
Formats the test api data frame to have date objects, turnaround times, adds the lab tags, etc.
}
\examples{
testApiFormat(All_Test)
All_Tests=testApiFormat(All_Tests)
}
