% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiIsPositive.R
\name{testApiIsPositive}
\alias{testApiIsPositive}
\title{Extracts Positive test}
\usage{
testApiIsPositive(DataFrame)
}
\arguments{
\item{DataFrame}{A Data Frame object from Test API}
}
\value{
NewFrame Contains only positive test
}
\description{
Extracts Positive test
}
\examples{
testApiIsPositive(PCR_Test)
testApiIsPositive(ByLab_Test)
}
