% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiMissingValues.R
\name{testApiMissingValues}
\alias{testApiMissingValues}
\title{Generates Plots that include dates in the x axis}
\usage{
testApiMissingValues(DataFrame)
}
\arguments{
\item{DataFrame}{A data frame that has already been manipulated by
testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}
}
\value{
A table with information on the number and fraction of missing values per variable
considered
}
\description{
Generates Plots that include dates in the x axis
}
\examples{
testApiMissingValues
testApiSubset(test_type="pcr") \%>\% testApiMissingValues()

}
