% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiMissingValuesCompact.R
\name{testApiMissingValuesCompact}
\alias{testApiMissingValuesCompact}
\title{Generates  Compact Table for Missing Values}
\usage{
testApiMissingValuesCompact(DataFrame)
}
\arguments{
\item{DataFrame}{A data frame that has already been manipulated by
testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}
}
\value{
A  compact table with information on the number and fraction of missing values per variable
considered
}
\description{
Generates  Compact Table for Missing Values
}
\examples{
testApiMissingValuesCompact()
testApiSubset(test_type="pcr") \%>\% testApiMissingValuesCompact()

}
