% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiPositivityPlot.R
\name{testApiPositivityPlot}
\alias{testApiPositivityPlot}
\title{Plots Positivity results}
\usage{
testApiPositivityPlot(DataFrame)
}
\arguments{
\item{DataFrame}{A Data Frame object with calculated positivity}
}
\value{
Positivity plot
}
\description{
Plots Positivity results
}
\examples{
testApiPositivityPlot(PCR_Test)
testApiPositivityPlot(ByLab_Test)
}
