% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiSubset.R
\name{testApiSubset}
\alias{testApiSubset}
\title{Subsets DataFrame by any of the available parameters in TestApi Table}
\usage{
testApiSubset(
  data_frame = All_Tests,
  sample_date_start = as.Date("2020-03-01"),
  sample_date_end = as.Date(Sys.Date()),
  report_date_start = as.Date("2020-03-01"),
  report_date_end = as.Date(Sys.Date()),
  upload_date_start = as.Date("2020-03-01"),
  upload_date_end = as.Date(Sys.Date()),
  test_type = "all",
  result_type = "all",
  lab_file = "all",
  lab_names = c("all"),
  lab_ids = c("all"),
  upload_method = "all",
  phone_numbers = c("all"),
  contact_type = "all"
)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}

\item{sample_date_start}{Default is set at as.Date("2020-03-01"). It's the date you want to start subsetting sample dates}

\item{sample_date_end}{Default is set at as.Date(Sys.Date()). It's the date you want to stop subsetting sample dates}

\item{report_date_start}{Default is set at as.Date("2020-03-01"). It's the date you want to start subsetting report dates}

\item{report_date_end}{Default is set at as.Date(Sys.Date()). It's the date you want to stop subsetting report dates}

\item{upload_date_start}{Default is set at as.Date("2020-03-01"). It's the date you want to start subsetting upload dates}

\item{upload_date_end}{Default is set at as.Date(Sys.Date()). It's the date you want to stop subsetting upload dates}

\item{test_type}{Default subsets by all types of test. Takes String value. Available test types are: Molecular, Serological, Antigens,Total Antibodies,Serological IgG Only}

\item{result_type}{Default subsets by all types of results. Takes String value. Available result types are:Negative, Not Detected, Inconclusive,Positive 2019-nCoV, Positive, Not Tested,Other,Positive IgM Only,Positive IgM and IgG,Positive IgG Only, Not Valid,Invalid, COVID-19 Negative, Presumptive Positive,"", COVID-19 Positive, SARS-CoV-2 Negative, SARS-CoV-2 Presumptive Positive,SARS-CoV-2 Positive}
}
\value{
NewFrame subseted Data Frame by the parameters specified
}
\description{
Subsets DataFrame by any of the available parameters in TestApi Table
}
\examples{
testApiSubset(PCR_Test)
testApiSubset(ByLab_Test)

}
