% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiDownload.R
\name{ApiDownload}
\alias{ApiDownload}
\title{Downloads the unique tests API from Bioportal for the selected date ranges}
\usage{
ApiDownload(
  email,
  password,
  from,
  to,
  api,
  timezone = "UTC",
  PR_Timezone = FALSE
)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}

\item{from}{Date (upload Date) from where to start downloading data.}

\item{to}{Date (upload date) to where to download the data (the last date is included)}

\item{api}{API to download}

\item{timezone}{timezone to use (defaults to "UTC")}

\item{PR_Timezone}{Boolean. If True, will set the timezone to "Etc/GMT+4"}
}
\value{
Selected API (use an API found in PilasPack::Endpoints) using your Bioportal credentials
}
\description{
Downloads the unique tests API from Bioportal for the selected date ranges
}
\examples{
DF=ApiDownload(your_email, your_password, start_date, end_date, api_to_use, PR_Timezone=TRUE)
All_Tests=ApiDownload(your_email, your_password, start_date, end_date, lag, PR_Timezone=TRUE)
Cases=ApiDownload(your_email, your_password, start_date, end_date, casesSummary, PR_Timezone=TRUE)
}
