% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsApiDownloadAll.R
\name{patientsApiDownloadAll}
\alias{patientsApiDownloadAll}
\title{Downloads the enteirity of the patients API}
\usage{
patientsApiDownloadAll(email, password)
}
\arguments{
\item{email}{Bioportal account email to be used to acces unique tests API}

\item{password}{Bioportal account password to be used to acces unique tests API}
}
\value{
data frame of all entities from entity API
}
\description{
Downloads the enteirity of the patients API
}
\examples{
patientsApiDownloadAll(email, password)
patients=entityApiDownloadAll(my_email, my_password)
}
