% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiCountDelays.R
\name{testApiCountDelays}
\alias{testApiCountDelays}
\title{Generates Plots that include dates in the x axis}
\usage{
testApiCountDelays(DataFrame = All_Tests)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn}
}
\value{
3 Table of distrubution of Turnaround times between: (1) Report Date and
Sample Date (2) Upload Date and Report Date (3) Upload Date and Sample Date
}
\description{
Generates Plots that include dates in the x axis
}
\examples{
testApiCountDelays(All_Tests)
testApiSubset(lab_tag="quest", test_type="pcr") \%>\% testApiCountDelays()


}
