% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testApiCountLabsAndType.R
\name{testApiCountLabsAndType}
\alias{testApiCountLabsAndType}
\title{Generates Table with number of tests by lab and type}
\usage{
testApiCountLabsAndType(data_frame = All_Tests)
}
\arguments{
\item{data_frame}{A data frame that has already been manipulated by
testApiCalculateTurnaround, testApiToDates and testApiMonthAsColumn.}
}
\value{
A table with information on the number of tests by lab and type
}
\description{
Generates Table with number of tests by lab and type
}
\examples{
testApiCountLabsAndType
testApiSubset(test_type="pcr") \%>\% testApiCountLabsAndType()

}
